requirements.txt
# Core dependencies - pinned versions for reliability
discord.py==2.3.2
httpx==0.25.0
colorama==0.4.6
aiohttp==3.9.0

# Build dependencies
pyinstaller==5.13.0

